@@ -9,8 +9,8 @@ package net.iatsoftware.website.forms;
  *
  * @author michael
  */
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
 
 public class EmailForm {
 
