@@ -14,9 +14,9 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.text.DateFormat;
 
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement(name="ServerErrorReport")
 @XmlAccessorType(XmlAccessType.NONE)
