@@ -16,10 +16,10 @@ import net.iatsoftware.website.generated.KeyType;
 
 import org.springframework.stereotype.Repository;
 
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Predicate;
-import javax.persistence.criteria.Root;
+import jakarta.persistence.criteria.CriteriaBuilder;
+import jakarta.persistence.criteria.CriteriaQuery;
+import jakarta.persistence.criteria.Predicate;
+import jakarta.persistence.criteria.Root;
 
 @Repository
 public class DefaultEncryptionKeyRepository extends GenericJpaRepository<Long, PartiallyEncryptedRSAKey> implements EncryptionKeyRepository {
@@ -33,7 +33,7 @@ public class DefaultEncryptionKeyRepository extends GenericJpaRepository<Long, P
         try {
             return this.entityManager.createQuery(query.select(root).where(pred)).getSingleResult();
         }
-        catch (javax.persistence.NoResultException | javax.persistence.NonUniqueResultException ex) {
+        catch (jakarta.persistence.NoResultException | jakarta.persistence.NonUniqueResultException ex) {
             return null;
         }
     }
