@@ -15,10 +15,10 @@ import net.iatsoftware.website.entities.IAT;
 
 import org.springframework.stereotype.Repository;
 
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Predicate;
-import javax.persistence.criteria.Root;
+import jakarta.persistence.criteria.CriteriaBuilder;
+import jakarta.persistence.criteria.CriteriaQuery;
+import jakarta.persistence.criteria.Predicate;
+import jakarta.persistence.criteria.Root;
 
 @Repository
 public class DefaultIATRepository extends GenericJpaRepository<Long, IAT> implements IATRepository {
@@ -41,7 +41,7 @@ public class DefaultIATRepository extends GenericJpaRepository<Long, IAT> implem
         try {
             return this.entityManager.createQuery(query.select(root).where(pred)).getSingleResult();
         }
-        catch (javax.persistence.NoResultException | javax.persistence.NonUniqueResultException ex) {
+        catch (jakarta.persistence.NoResultException | jakarta.persistence.NonUniqueResultException ex) {
             return null;
         }
     }
