@@ -14,10 +14,10 @@ import net.iatsoftware.website.entities.User;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
-import javax.persistence.criteria.Predicate;
+import jakarta.persistence.criteria.CriteriaBuilder;
+import jakarta.persistence.criteria.CriteriaQuery;
+import jakarta.persistence.criteria.Root;
+import jakarta.persistence.criteria.Predicate;
 
 @Repository
 public class DefaultUserRepository extends GenericJpaRepository<Long, User> implements UserRepository {
