@@ -21,7 +21,7 @@ import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Service;
 import org.thymeleaf.context.Context;
-import org.thymeleaf.spring5.SpringTemplateEngine;
+import org.thymeleaf.spring6.SpringTemplateEngine;
 import org.thymeleaf.templatemode.TemplateMode;
 import org.thymeleaf.templateresolver.ClassLoaderTemplateResolver;
 import org.thymeleaf.templateresolver.ITemplateResolver;
@@ -32,7 +32,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
 import javax.inject.Inject;
-import javax.mail.internet.MimeMessage;
+import jakarta.mail.internet.MimeMessage;
 
 @Service("MailService")
 @PropertySource("classpath:environment.properties")
@@ -48,7 +48,7 @@ public class DefaultMailService implements MailService {
     JavaMailSender mailSender;
 
     @Override
-    public void sendEmail(EmailParameters params) throws javax.mail.MessagingException {
+    public void sendEmail(EmailParameters params) throws jakarta.mail.MessagingException {
         try {
             final Context ctx = new Context();
             ctx.setVariables(params.getParameters());
@@ -64,7 +64,7 @@ public class DefaultMailService implements MailService {
                     helper.addInline(trip.getFirst(), new ClassPathResource(trip.getSecond()),
                             "image/" + trip.getSecond().substring(trip.getSecond().lastIndexOf('.')));
                     ctx.setVariable(trip.getFirst(), trip.getThird());
-                } catch (javax.mail.MessagingException ex) {
+                } catch (jakarta.mail.MessagingException ex) {
                     logger.error("Error embedding inline image " + trip.getFirst() + " in email", ex);
                 }
             });
