@@ -11,8 +11,8 @@ package net.iatsoftware.website.validation;
  */
 
 import java.net.URL;
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 
 public class DomainValidator implements ConstraintValidator<DomainValid, CharSequence> {
     
