@@ -13,8 +13,8 @@ package net.iatsoftware.website.validation;
 import net.iatsoftware.website.repositories.RepositoryManager;
 
 import javax.inject.Inject;
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 
 public class RegisteredEmailValidator implements ConstraintValidator<RegisteredEmailValid, CharSequence> {
     @Inject RepositoryManager repositoryManager;
