@@ -15,8 +15,8 @@ import net.iatsoftware.website.entities.Client;
 import net.iatsoftware.website.repositories.RepositoryManager;
 
 import javax.inject.Inject;
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 
 public class TestNameValidator implements ConstraintValidator<TestNameValid, OauthRegistrationForm> {
     
