@@ -1,14 +1,14 @@
 //
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
-// See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
+// This file was generated by the Eclipse Implementation of JAXB, v3.0.2 
+// See https://eclipse-ee4j.github.io/jaxb-ri 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2023.07.15 at 01:54:43 AM EDT 
+// Generated on: 2023.08.01 at 01:11:26 AM EDT 
 //
 
 
 package net.iatsoftware.website.generated;
 
-import javax.xml.bind.annotation.XmlRegistry;
+import jakarta.xml.bind.annotation.XmlRegistry;
 
 
 /**
@@ -45,11 +45,11 @@ public class ObjectFactory {
     }
 
     /**
-     * Create an instance of {@link net.iatsoftware.website.generated.ServerErrorReport }
+     * Create an instance of {@link ServerErrorReport }
      * 
      */
-    public net.iatsoftware.website.generated.ServerErrorReport createServerErrorReport() {
-        return new net.iatsoftware.website.pojos.ServerErrorReport();
+    public ServerErrorReport createServerErrorReport() {
+        return new ServerErrorReport();
     }
 
 }
