@@ -1,8 +1,8 @@
 //
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
-// See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
+// This file was generated by the Eclipse Implementation of JAXB, v3.0.2 
+// See https://eclipse-ee4j.github.io/jaxb-ri 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2023.07.15 at 01:54:43 AM EDT 
+// Generated on: 2023.08.01 at 01:11:26 AM EDT 
 //
 
 
@@ -10,11 +10,10 @@ package net.iatsoftware.website.generated;
 
 import java.util.ArrayList;
 import java.util.List;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlType;
-import net.iatsoftware.website.pojos.SerializableException;
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlType;
 
 
 /**
@@ -48,9 +47,9 @@ public class ServerErrorReport {
 
     @XmlElement(name = "Timestamp", required = true)
     protected String timestamp;
-    @XmlElement(name = "Exception", required = true, type = SerializableException.class)
-    protected SerializableException exception;
-    @XmlElement(name = "InnerException", type = SerializableException.class)
+    @XmlElement(name = "Exception", required = true)
+    protected Exception exception;
+    @XmlElement(name = "InnerException")
     protected List<Exception> innerException;
 
     /**
@@ -98,7 +97,7 @@ public class ServerErrorReport {
      *     
      */
     public void setException(Exception value) {
-        this.exception = ((SerializableException) value);
+        this.exception = value;
     }
 
     /**
@@ -107,7 +106,7 @@ public class ServerErrorReport {
      * <p>
      * This accessor method returns a reference to the live list,
      * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
+     * returned list will be present inside the Jakarta XML Binding object.
      * This is why there is not a <CODE>set</CODE> method for the innerException property.
      * 
      * <p>
